+++ src/analysis/type_precheck.cpp
@@ -224,7 +224,7 @@ DEFINE_AST_VISITOR_IMPL(TypePrecheck, StructDecl) {
 
     if (m->get_node_type() == ASTNodeType::VAR_DECL) { /// member variable without initial value
       str name = ast_cast<VarDecl>(m)->get_name();
-      p->set_member_index(name, i);
+      p->set_member_index(name, (int)i);
       (*ty)[i] = m->get_type();
     } else if (m->get_node_type() == ASTNodeType::ASSIGN) { /// member variable with an initial value
       auto bm = ast_cast<Assignment>(m);
@@ -235,12 +235,12 @@ DEFINE_AST_VISITOR_IMPL(TypePrecheck, StructDecl) {
       auto decl = ast_cast<VarDecl>(bm->get_lhs());
 
       (*ty)[i] = decl->get_type();
-      p->set_member_index(decl->get_name(), i);
+      p->set_member_index(decl->get_name(), (int)i);
     } else if (m->get_node_type() == ASTNodeType::FUNC_DECL) { /// member functions
       auto f = ast_cast<FunctionDecl>(m);
 
       (*ty)[i] = f->get_type();
-      p->set_member_index(f->get_name(), i);
+      p->set_member_index(f->get_name(), (int)i);
     } else {
       error(p, "Invalid struct member");
     }
